Introduce GIT_REF environment variable in GitHub workflow
This commit refactors the way the GIT_REF variable is managed in our GitHub workflows. Previously, it was directly assigned inside the Safety mutation tests script. However, for better flexibility and readability, we decided to move this variable into the GitHub workflow environment variables. This approach allows us to control this parameter in a more centralized manner. Additionally, the default value for GIT_REF in the mutation.sh script has been updated to 'origin/master-new'.